+++ scripts/Groovy/CsvQuery.groovy
@@ -25,6 +25,8 @@ System.setProperty('java.util.logging.SimpleFormatter.format',
 long startTime = System.currentTimeMillis()
 def cli = new CliBuilder(usage: 'groovy CsvQuery.groovy [options]')
 cli.i(longOpt: 'input', args: 1, argName: 'input', 'Input CSV file')
+cli.is(longOpt: 'input-file-separator', args: 1, argName: 'input-file-separator', defaultValue: ',',
+        'The separator character used in the input file. Defaults to ,')
 cli.o(longOpt: 'output', args: 1, argName: 'output', 'Output CSV file')
 cli.q(longOpt: 'query', args: 1, argName: 'query', required: true, 'Yaml file containing the query')
 cli.t(longOpt: 'threads', args: 1, argName: 'threads', defaultValue: '3', 'Number of threads to use for processing')
@@ -37,6 +39,7 @@ if (!options) {
 CsvListReader csvReader
 CsvMapWriter csvWriter
 
+char inputSeparator = options.is
 Query query = QueryUtil.parseQuery(new File(options.q).text)
 String inputFile = options.i ?: query.from
 assert inputFile : 'Input file name not specified'
@@ -47,9 +50,12 @@ AtomicInteger recordCount = new AtomicInteger(0)
 int threads = Integer.parseInt(options.t)
 ThreadPoolExecutor executor = new ThreadPoolExecutor(1, threads, 0L, TimeUnit.MILLISECONDS,
         new LinkedBlockingQueue<>(threads * 5), new ThreadPoolExecutor.CallerRunsPolicy())
+CsvPreference inputFilePref = new CsvPreference.Builder(
+        CsvPreference.STANDARD_PREFERENCE.quoteChar, (int)inputSeparator, CsvPreference.STANDARD_PREFERENCE.endOfLineSymbols)
+        .build()
 
 try {
-    csvReader = new CsvListReader(new FileReader(inputFile), CsvPreference.STANDARD_PREFERENCE)
+    csvReader = new CsvListReader(new FileReader(inputFile), inputFilePref)
     csvWriter = new CsvMapWriter(writer, CsvPreference.STANDARD_PREFERENCE)
     AsyncCsvWriter asyncWriter = new AsyncCsvWriter(csvWriter)
     String[] outputHeaders